+++ src/core/agents.jl
@@ -103,6 +103,18 @@ agent = Person(1, (1, 1), 40, 2000)
 agent.moneyz = 1000
 agent.age = 20 # this throws an error
 ```
+Notice that you can also use default values for some fields, in this case you 
+will need to specify the field names with the non-default values
+```julia
+@agent Person{T} GridAgent{2} begin
+    age::Int = 30
+    moneyz::T
+end
+# default age value
+Person(id = 1, pos = (1, 1), moneyz = 2000)
+# new age value
+Person(1, (1, 1), 40, 2000)
+```
 ### Example with optional hierarchy
 An alternative way to make the above structs, that also establishes
 a user-specific subtyping hierarchy would be to do:
@@ -219,7 +231,7 @@ macro agent(new_name, base_type, super_type, extra_fields)
             expr = quote
                 # Also notice that we escape supertype and interpolate it twice
                 # because this is expected to already be defined in the calling module
-                mutable struct $name <: $$(esc(super_type))
+                @kwdef mutable struct $name <: $$(esc(super_type))
                     $(base_fields...)
                     $(additional_fields...)
                 end