+++ src/simulations/sample.jl
@@ -39,7 +39,6 @@ end
 function add_newids!(model, org_ids, newids)
     # `counter` counts the number of occurencies for each item, it comes from DataStructure.jl
     count_newids = counter(newids)
-    n = nextid(model)
     for id in org_ids
         noccurances = count_newids[id]
         agent = model[id]
@@ -47,10 +46,7 @@ function add_newids!(model, org_ids, newids)
             remove_agent!(agent, model)
         else
             for _ in 2:noccurances
-                newagent = deepcopy(agent)
-                newagent.id = n
-                add_agent_pos!(newagent, model)
-                n += 1
+                replicate!(agent, model)
             end
         end
     end
@@ -59,9 +55,9 @@ end
 """
     replicate!(agent, model; kwargs...) 
 
-Create a new agent at the same position of the given agent, copying the values
-of its fields. With the `kwargs` it is possible to override the values by specifying
-new ones for some fields. 
+Add a new agent to the `model` at the same position of the given agent, copying
+the values of its fields. With the `kwargs` it is possible to override the values
+by specifying new ones for some fields. 
 Return the new agent instance.
 
 ## Example
@@ -77,16 +73,25 @@ a = A(1, (2, 2), 0.5, 0.5)
 b = replicate!(a, model; w = 0.8)
 ```
 """
-function replicate!(agent, model; kwargs...)
-    newagent = deepcopy(agent)
-    for (key, value) in kwargs
-        setfield!(newagent, key, value)
-    end
-    newagent.id = nextid(model)
+function replicate!(agent::A, model; kwargs...) where {A<:AbstractAgent}
+    args = new_args(agent::A, model; kwargs...) 
+    newagent = A(nextid(model), args...)
     add_agent_pos!(newagent, model)
     return newagent
 end
 
-function Base.deepcopy(agent::A) where {A<:AbstractAgent}
-    return A((deepcopy(getfield(agent, name)) for name in fieldnames(A))...)
+function new_args(agent::A, model; kwargs...) where {A<:AbstractAgent}
+    fields = fieldnames(A)
+    idx_id = findfirst(x -> x == :id, fields)
+    fields_no_id = tuple(fields[1:idx_id-1]..., fields[idx_id+1:end]...)
+    if isempty(kwargs)
+        new_args = (deepcopy(getfield(agent, x)) for x in fields_no_id)
+    else
+        kwargs_nt = NamedTuple(kwargs)
+        new_args = (choose_arg(x, kwargs_nt, agent) for x in fields_no_id)
+    end
+end
+
+function choose_arg(x, kwargs_nt, agent)
+    return deepcopy(getfield(hasproperty(kwargs_nt, x) ? kwargs_nt : agent, x))
 end